requirements.txt
# ── FastAPI & server ─────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1

# ── Auth ─────────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ── LangChain + Azure OpenAI ─────────────────────────────────────────────────
langchain==0.2.1
langchain-core==0.2.1
langchain-openai==0.1.7
langchain-community==0.2.1
openai==1.30.1

# ── Azure SDKs ───────────────────────────────────────────────────────────────
azure-search-documents==11.6.0b4
azure-core==1.30.1
azure-storage-blob==12.20.0

# ── RAG / document ingestion ─────────────────────────────────────────────────
pdfplumber==0.11.0
pandas==2.2.2
openpyxl==3.1.2
tiktoken==0.7.0

# ── Database ─────────────────────────────────────────────────────────────────
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# ── Redis (async) ────────────────────────────────────────────────────────────
redis[asyncio]==5.0.4

# ── Monitoring / telemetry ───────────────────────────────────────────────────
opencensus-ext-azure==1.1.13
opencensus==0.11.4

# ── Utilities ────────────────────────────────────────────────────────────────
httpx==0.27.0
python-dotenv==1.0.1
tenacity==8.3.0
